added simple notification of reminder triggers

+++ app/src/main/kotlin/kr/ac/snu/hcil/omnitrack/ui/pages/trigger/ATriggerViewHolder.kt
@@ -69,7 +69,7 @@ abstract class ATriggerViewHolder<T : OTTrigger>(parent: ViewGroup, val listener
         applyTriggerStateToView()
     }
 
-    private val onTriggerFired: ((Any, Int) -> Unit) = {
+    private val onTriggerFired: ((Any, Long) -> Unit) = {
         sender, triggeredTime ->
         applyTriggerStateToView()
     }
